# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "XR50TrainingAssetRepo",
    "version": "1.0"
  },
  "paths": {
    "/xr50/TrainingAssetRepository/asset_management": {
      "get": {
        "tags": [
          "asset_management"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Asset"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "asset_management"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "TenantName": {
                    "type": "string"
                  },
                  "FileName": {
                    "type": "string"
                  },
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "TenantName": {
                  "style": "form"
                },
                "FileName": {
                  "style": "form"
                },
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "asset_management"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "TenantName": {
                    "type": "string"
                  },
                  "ParentId": {
                    "type": "string"
                  },
                  "Type": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "TenantName": {
                  "style": "form"
                },
                "ParentId": {
                  "style": "form"
                },
                "Type": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/asset_management/{id}": {
      "get": {
        "tags": [
          "asset_management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "asset_management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/asset_management/share": {
      "get": {
        "tags": [
          "asset_management"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Share"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Share"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Share"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "asset_management"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Share"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Share"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Share"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Share"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Share"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Share"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/asset_management/share/{id}": {
      "put": {
        "tags": [
          "asset_management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Share"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Share"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Share"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "asset_management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/asset_management/directory": {
      "post": {
        "tags": [
          "asset_management"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwncloudDirectory"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OwncloudDirectory"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OwncloudDirectory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Asset": {
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "shareList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "materialList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OwncloudDirectory": {
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "owncloudPath": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Share": {
        "type": "object",
        "properties": {
          "shareId": {
            "type": "string",
            "nullable": true
          },
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "fileId": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ShareType"
          },
          "target": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ShareType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  }
}